+++ src/main/java/com/generation/blogpessoal/repository/TemaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface TemaRepository extends JpaRepository<Tema, Integer> {
+public interface TemaRepository extends JpaRepository<Tema, Long> {
 
     public List<Tema> findAllByDescricaoContainingIgnoreCase(@Param("descricao") String descricao);
 }
